Chapter 11: fix bug
Didn't remove initialization of u, v, and w in the camera constructor.
This meant the class variable version of these were never initialized,
and so in get_ray(), they have random values, so the output is
meaningless.
Interestingly enough, for small apartures, the image looks the same
as it did before adding defocus blur, but at larger apertures, it
would become a uniform colour.
In any case, removing the leftover initialization of local variables
u, v, and w in the constructor fixed this right up.